combine webpack multi configs.
