+++ webpack.dev.js
@@ -1,7 +1,9 @@
 const merge = require('webpack-merge');
-const prod_config = require('./webpack.config.js');
+const prod_configs = require('./webpack.config.js');
 
-module.exports = merge.merge(prod_config, {
-  mode: 'development',
-  devtool: 'inline-source-map',
-});
+module.exports = prod_configs.map((config) =>
+  merge.merge(prod_configs[0], {
+    mode: 'development',
+    devtool: 'inline-source-map',
+  })
+);
