--- webpack.web.js
@@ -1,58 +1,0 @@
-//@ts-check
-
-'use strict';
-
-const path = require('path');
-const webpack = require('webpack');
-
-/**@type {import('webpack').Configuration}*/
-const nodelessConfig = {
-  target: 'webworker', // vscode extensions run in a Node.js-context 📖 -> https://webpack.js.org/configuration/node/
-
-  mode: 'development',
-
-  entry: './extensionWeb.ts', // the entry point of this extension, 📖 -> https://webpack.js.org/configuration/entry-context/
-  output: {
-    // the bundle is stored in the 'out' folder (check package.json), 📖 -> https://webpack.js.org/configuration/output/
-    path: path.resolve(__dirname, 'out'),
-    filename: 'extensionWeb.js',
-    libraryTarget: 'umd',
-  },
-  devtool: 'inline-source-map',
-  externals: {
-    vscode: 'commonjs vscode', // the vscode-module is created on-the-fly and must be excluded. Add other modules that cannot be webpack'ed, 📖 -> https://webpack.js.org/configuration/externals/
-  },
-  resolve: {
-    // support reading TypeScript and JavaScript files, 📖 -> https://github.com/TypeStrong/ts-loader
-    extensions: ['.ts', '.js'],
-    alias: {
-      platform: path.resolve(__dirname, 'src', 'platform', 'browser'),
-    },
-  },
-  module: {
-    rules: [
-      {
-        test: /\.ts$/,
-        exclude: /node_modules/,
-        use: [
-          {
-            loader: 'ts-loader',
-          },
-        ],
-      },
-    ],
-  },
-  plugins: [
-    new webpack.IgnorePlugin({
-      resourceRegExp: /\/neovim$/,
-    }),
-    new webpack.IgnorePlugin({
-      resourceRegExp: /\/imswitcher$/,
-    }),
-    new webpack.IgnorePlugin({
-      resourceRegExp: /child_process$/,
-    }),
-  ],
-};
-
-module.exports = nodelessConfig;
